Add process metrics importer
Move the logic for importing process metrics from the Prometheus broker
to public API. This allows users to get access to the process metrics
without having to load the HTTP server for Prometheus scraping.
